requirements.txt
coverage==4.4.2
coveralls==1.2.0
docutils==0.14
Django==2.0.1
djangorestframework==3.7.7
django-nose==1.4.5